Now we have a standard way to call observers
Instead of having different methods to call observers (like update, and
displayStuff) all you need to do now is just call your reference to the
object.

Next step is to keep the list of observers in one place. That way, when
we need to add another observer, instead writing it into the lambda,
we'll just add it to a list.

(obeserver pattern observation pattern)