Defined methods that were only declared.
factory::Pizza::~Pizza was declared but never defined.
factory::CheezePizza::cost was declared but never defined.

----

The cmake error message was:

CMakeFiles/factoryexe.dir/Factory.cpp.o:/home/matthew/projects/lyase/joggingthecpppath/Factory.cpp:function
factory::CheezePizza::CheezePizza(): error: undefined reference to
'vtable for factory::CheezePizza'
factory/libfactory.a(Pizza.cpp.o):/home/matthew/projects/lyase/joggingthecpppath/factory/Pizza.cpp:function
factory::Pizza::Pizza(): error: undefined reference to 'vtable for
factory::Pizza'

----

So from the first line we can see that:

 -- undefined reference to 'vtable for factory::CheezePizza'

means, the vtable for factory::CheezePizza is declared (and referenced by someone), but
not defined. The vtable is needed to make a class virtual. So this is
saying, I want to make this class virtual but something's missing.

The next line says the same thing about the factory::Pizza class.

I think what we can infer from this is that there is a method that is
declared as virtual, in Pizza, but never defined.

Having a look it appears that that method was the destructor.

Fixed..